拆分 SFTP 文件操作流程
